Space X Falcon 9 First Stage Landing Prediction
Space X advertises Falcon 9 rocket launches on its website with a cost of 62 million dollars; other providers cost upward of 165 milliondollars each, much of the savings is because Space X can reuse the first stage. Therefore if we can determine if the first stage will land, wecan determine the cost of a launch. This information can be used if an alternate company wants to bid against space X for a rocket launch. Inthis lab, you will create a machine learning pipeline to predict if the first stage will land given the data from the preceding labs.